So #11361 prompted this PR -- it's been on my plate for a while. Recent work from my team (not on OpenSSL) tells our guts this is a better approach. Tagging @acaldaya for a sanity check.

It also recovers a tiny bit of the performance regression mentioned in #6512. Limited evidence below.

NB the actual curve math BN arithmetic parts are auto generated from tooling.

IMO master and 1.1.1.

Edit: Oh and I did not test `no-asm` which I think is important here, since all the arithmetic without that is handled by `ecp_mont` and not `ecp_smpl`.

Side note: this took me longer than I'd like to admit -- too long away from the OpenSSL codebase! I'll try to be more active.

## Before

```
bbrumley:~/svnrepos/openssl$ apps/openssl speed ecdhp384
Doing 384 bits  ecdh's for 10s: 9437 384-bits ECDH ops in 10.00s
version: 3.0.0-dev
built on: built on: Sat Mar 28 18:29:00 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_BUILDING_OPENSSL -DOPENSSL_PIC -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              op      op/s
 384 bits ecdh (nistp384)   0.0011s    943.7
bbrumley:~/svnrepos/openssl$ apps/openssl speed ecdsap384
Doing 384 bits sign ecdsa's for 10s: 8733 384 bits ECDSA signs in 10.00s 
Doing 384 bits verify ecdsa's for 10s: 11723 384 bits ECDSA verify in 10.00s
version: 3.0.0-dev
built on: built on: Sat Mar 28 18:29:00 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_BUILDING_OPENSSL -DOPENSSL_PIC -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              sign    verify    sign/s verify/s
 384 bits ecdsa (nistp384)   0.0011s   0.0009s    873.3   1172.3
```

## After

```
bbrumley:~/svnrepos/openssl$ apps/openssl speed ecdhp384
Doing 384 bits  ecdh's for 10s: 10183 384-bits ECDH ops in 10.01s
version: 3.0.0-dev
built on: built on: Sat Mar 28 18:36:26 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_BUILDING_OPENSSL -DOPENSSL_PIC -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              op      op/s
 384 bits ecdh (nistp384)   0.0010s   1017.3
bbrumley:~/svnrepos/openssl$ apps/openssl speed ecdsap384
Doing 384 bits sign ecdsa's for 10s: 9678 384 bits ECDSA signs in 10.00s 
Doing 384 bits verify ecdsa's for 10s: 12008 384 bits ECDSA verify in 10.00s
version: 3.0.0-dev
built on: built on: Sat Mar 28 18:36:26 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_BUILDING_OPENSSL -DOPENSSL_PIC -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              sign    verify    sign/s verify/s
 384 bits ecdsa (nistp384)   0.0010s   0.0008s    967.8   1200.8
```

## Machine

```
Architecture:          x86_64
CPU op-mode(s):        32-bit, 64-bit
Byte Order:            Little Endian
CPU(s):                48
On-line CPU(s) list:   0-47
Thread(s) per core:    2
Core(s) per socket:    12
Socket(s):             2
NUMA node(s):          2
Vendor ID:             GenuineIntel
CPU family:            6
Model:                 85
Model name:            Intel(R) Xeon(R) Silver 4116 CPU @ 2.10GHz
Stepping:              4
CPU MHz:               800.337
CPU max MHz:           3000.0000
CPU min MHz:           800.0000
BogoMIPS:              4201.70
Virtualization:        VT-x
L1d cache:             32K
L1i cache:             32K
L2 cache:              1024K
L3 cache:              16896K
NUMA node0 CPU(s):     0-11,24-35
NUMA node1 CPU(s):     12-23,36-47
Flags:                 fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid dca sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb cat_l3 cdp_l3 invpcid_single pti intel_ppin ssbd mba ibrs ibpb stibp tpr_shadow vnmi flexpriority ept vpid fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm cqm mpx rdt_a avx512f avx512dq rdseed adx smap clflushopt clwb intel_pt avx512cd avx512bw avx512vl xsaveopt xsavec xgetbv1 xsaves cqm_llc cqm_occup_llc cqm_mbm_total cqm_mbm_local dtherm ida arat pln pts hwp hwp_act_window hwp_pkg_req pku ospke flush_l1d
```