@beldmit 2020-03-29
The in-house benchmarking for GOST using 1.1.1 branch gives me +10% speedup on signature (1220 vs 1110) and minor speedup on verification (1470 vs 1440). 

BTW, 1.1.1 does not have `BN_priv_rand_range_ex` function