@romen 2020-03-29
> BTW, 1.1.1 does not have `BN_priv_rand_range_ex` function

To approve for backport we need to address this too.

I would personally approve it for backport as this is technically a security fix, not just a performance improvement: independent randomization of coordinates for the two accumulator points prevent certain attacks.

Maybe @mattcaswell can share his view on the matter of backporting to 1.1.1