@kroeckx 2020-03-29
If you consider this a security fix, I'm happy to backport this to 1.1.1.