@romen 2020-03-29
> If you consider this a security fix, I'm happy to backport this to 1.1.1.

I mentioned @mattcaswell but of course @kroeckx opinion would also work, what I'd like to have is an approval for backport to 1.1.1 from an OMC&OTC member.

My opinion is that this would qualify as a security fix, with the added bonus of nice performance gain and keeping in sync this codebase between `master` and 1.1.1, so I am favorable to backport to 1.1.1.