@romen 2020-03-29
> The in-house benchmarking for GOST using 1.1.1 branch gives me +10% speedup on signature (1220 vs 1110) and minor speedup on verification (1470 vs 1440).
> 
> **Update:** a bit less, 1220 vs 1180 for signature, 1470 vs 1460 for verification.

These changes should not have an effect on verification, as verification uses the fast&insecure path as no input there is secret: the variance you are seeing is just regular sampling variance.

You should be able to appreciate performance gains benchmarking signature generation, key generation and the second half of DH (private scalar times the peer public point).