@bbbrumley 2020-03-30
> Do you have a paper on the math?

I wish we did :\ It's a long story, that started with the ladder work just before 1.1.1. The academic metric for ECC formula efficiency is GF multiplies and squares. We noticed within the OpenSSL codebase (and I'm sure elsewhere) this is not a great metric (under estimates the cost of "other stuff" e.g. GF adds and other CRUD, I don't even know what). We were left with 2 options

1. Model it
2. Just measure it

We chose (2) and wrote a tool that takes [formulae from EFD](https://www.hyperelliptic.org/EFD/) and spits out OpenSSL code. It works pretty OK (and we used it for both ecp and ec2m ladder functionality at the time).

We were hoping to use it as part of a bigger overhaul of ec2m in OpenSSL. But it turns out nothing useful has actually happened in ec2m efficiency in like 20 years, despite what academia says. (With respect to the measurements in OpenSSL, at least.)

So that paper never happened. I have hopes to revisit at a later date (not ec2m focused, but more generally), but no guarantees.