@bernd-edlinger 2020-03-30
Ah, thanks anyway, I have written a somewhat nice AES constant time code
that works without assembler (in january already, but it stalled in the review process,
I am kind of used to that)  I think it's math has a lot in common with the GF fields you
use in this PR, maybe you or @romen find time to give it a review #11411 
I would like to see that one in 1.1.1f but I feel a bit exhausted.