@bernd-edlinger 2020-03-30
> happened in ec2m efficiency

I can only say that ec2m is soooooooo sloow.