@bernd-edlinger 2020-03-30
I started to build always with no-ec2m (that way the final 1.1.0 broke for me)