@romen 2020-03-30
@bernd-edlinger even if the last changes were mostly cosmetic and about code comments, I believe you need to formally reconfirm your approval.