@romen 2020-03-30
I marked this as `approval: ready` (at least regarding `master`) and I am currently checking what would be required to backport to 1.1.1 to see if the changes are trivial enough or if we need a separate PR.