@kroeckx 2020-03-30
On Mon, Mar 30, 2020 at 10:42:55AM -0700, Nicola Tuveri wrote:
> This is the extent of the patching needed to get this to work in 1.1.1:
> 
> ```diff
> diff --git a/crypto/ec/ecp_smpl.c b/crypto/ec/ecp_smpl.c
> index 2eed64d054..0feb88b6e1 100644
> --- a/crypto/ec/ecp_smpl.c
> +++ b/crypto/ec/ecp_smpl.c
> @@ -1508,13 +1508,13 @@ int ec_GFp_simple_ladder_pre(const EC_GROUP *group,
>  
>      /* make sure lambda (r->Y here for storage) is not zero */
>      do {
> -        if (!BN_priv_rand_range_ex(r->Y, group->field, ctx))
> +        if (!BN_priv_rand_range(r->Y, group->field))
>              return 0;
>      } while (BN_is_zero(r->Y));
>  
>      /* make sure lambda (s->Z here for storage) is not zero */
>      do {
> -        if (!BN_priv_rand_range_ex(s->Z, group->field, ctx))
> +        if (!BN_priv_rand_range(s->Z, group->field))
>              return 0;
>      } while (BN_is_zero(s->Z));
>  
> ```
> 
> For me this qualifies as trivial: opinions? @kroeckx

I'm happy with that as change for 1.1.1 compared to master.


Kurt