@bernd-edlinger 2020-04-01
OH SORRY!
Please disregard my previous comment, there are no objections for _this_ PR from my side.
That was meant for the other PR that introduce blinding in wNAF #11439