@bernd-edlinger 2020-04-01
I would even say, my approval also holds for the merge to 1.1.1 where
Nicola pointed out how to fix the merge commit manually.