@romen 2020-04-01
Merged to `master` with:
- 09736245b1 [test] Make sm2_internal_test less fragile to changes in the ec module
- a4a93bbfb0 [crypto/ec] Ladder tweaks

Merged to `1.1.1` with:
- 3c61ea3671 [test] Make sm2_internal_test less fragile to changes in the ec module
- d0771a9fdb [crypto/ec] Ladder tweaks